Move aalloc() et al to new hdr/src c4/alloc.hpp
