+++ test/CMakeLists.txt
@@ -43,6 +43,7 @@ c4core_test(error            test_error.cpp)
 c4core_test(error_exception  test_error_exception.cpp)
 c4core_test(blob             test_blob.cpp)
 c4core_test(memory_util      test_memory_util.cpp)
+c4core_test(alloc            test_alloc.cpp)
 c4core_test(memory_resource  test_memory_resource.cpp)
 c4core_test(allocator        test_allocator.cpp)
 c4core_test(ctor_dtor        test_ctor_dtor.cpp)
