+++ tools/amalgamate.py
@@ -77,6 +77,7 @@ INSTRUCTIONS:
         am.hdrfile("src/c4/ext/debugbreak/debugbreak.h", "c4/ext/debugbreak/debugbreak.h", "DEBUG_BREAK_H"),
         "src/c4/error.hpp",
         "src/c4/memory_util.hpp",
+        "src/c4/alloc.hpp",
         "src/c4/memory_resource.hpp",
         "src/c4/ctor_dtor.hpp",
         "src/c4/allocator.hpp",
@@ -120,6 +121,7 @@ INSTRUCTIONS:
         "src/c4/format.cpp",
         "src/c4/memory_util.cpp",
         "src/c4/char_traits.cpp",
+        "src/c4/alloc.cpp",
         "src/c4/memory_resource.cpp",
         "src/c4/utf.cpp",
         "src/c4/base64.cpp",
